:sparkles: Carrito y sesiones de usuario terminado

+++ app/Http/Controllers/CartController.php
@@ -13,49 +13,36 @@ class CartController extends Controller
 {
     public function index()
     {
-        // Primero buscamos por el token de la cookie si existe
-        $token = request()->cookie('cart_token');
-        $cart = null;
-
-        if ($token) {
-            $cart = Cart::with('items.product.images')
-                ->where('token', $token)
-                ->first();
-        }
-
-        // Si el usuario está autenticado y no tenemos un carrito de token, buscamos el del usuario
-        if (Auth::check() && !$cart) {
-            $cart = Cart::with('items.product.images')
-                ->where('customer_id', Auth::id())
-                ->first();
-        }
-        return view('cart_items', compact('cart'));
+        $cart = $this->getCart();
+        return view('cart', compact('cart'));
     }
 
     public function add(Request $request, $productId)
     {
-        $cart = $this->getCart();
-        $product = Product::findOrFail($productId);
+        try {
+            $cart = $this->getCart();
+            $product = Product::findOrFail($productId);
 
-        // Buscar si el producto ya está en el carrito
-        $cartItem = CartItem::where('cart_id', $cart->id)
-            ->where('product_id', $product->id)
-            ->first();
+            // Verificar si el producto ya está en el carrito
+            $cartItem = $cart->items()->where('product_id', $product->id)->first();
 
-        if ($cartItem) {
-            // Si ya está, aumentar la cantidad
-            $cartItem->quantity += 1;
-        } else {
-            // Si no está, agregarlo al carrito
-            $cartItem = new CartItem(['product_id' => $product->id, 'quantity' => 1]);
-            $cart->items()->save($cartItem);
+            if ($cartItem) {
+                // Si ya está, aumentar la cantidad
+                $cartItem->quantity += 1;
+            } else {
+                // Si no está, agregarlo al carrito
+                $cartItem = new CartItem(['product_id' => $product->id, 'quantity' => 1]);
+                $cart->items()->save($cartItem);
+            }
+
+            $cartItem->save();
+
+            // Devolver la respuesta con la cantidad de productos en el carrito
+            $cartItemCount = $cart->items->sum('quantity');
+            return response()->json(['cartItemCount' => $cartItemCount]);
+        } catch (\Exception $e) {
+            return response()->json(['message' => 'Error al agregar el producto: ' . $e->getMessage()], 500);
         }
-
-        $cartItem->save();
-        // Devolvemos la respuesta con la cantidad de productos en el carrito
-        $cartItemCount = $cart->items->sum('quantity');
-
-        return response()->json(['cartItemCount' => $cartItemCount]);
     }
 
     public function remove(Request $request, $itemId)
@@ -66,69 +53,61 @@ class CartController extends Controller
             $cart->items()->where('id', $itemId)->delete();
         }
 
-        // Devolvemos la respuesta con un mensaje de éxito y la cantidad de productos en el carrito
-        if ($request->ajax()) {
-            return response()->json(['success' => true, 'message' => 'Producto eliminado del carrito', 'cartItemCount' => $cart->items->sum('quantity')]);
-        }
-        return redirect()->route('cart')->with('success', 'Producto eliminado del carrito');
+        // Después de eliminar, recalcula el contador del carrito
+        $cartItemCount = $cart ? $cart->items->sum('quantity') : 0;
+        return response()->json(['cartItemCount' => $cartItemCount]);
     }
 
     public function clear()
     {
-        $cart = Cart::where('token', request()->cookie('cart_token'))->orWhere('customer_id', Auth::id())->first(); // Buscar el carrito por token o por usuario autenticado
+        $cart = $this->getCart();
 
-        // Eliminar todos los elementos del carrito
         if ($cart) {
             $cart->items()->delete();
         }
-        return redirect()->route('cart')->with('success', 'Carrito vaciado con éxitDo');
+        return redirect()->route('cart')->with('success', 'Carrito vaciado con éxito');
     }
 
     private function getCart()
     {
-        // Obtener el token del carrito desde la cookie o generar un nuevo token si no existe
-        $token = request()->cookie('cart_token') ?: Str::uuid();
-
         if (Auth::check()) {
-            // Si el usuario está autenticado, buscamos o creamos un carrito para él
+            // Usuario autenticado: buscar o crear un carrito para él
             $cart = Cart::firstOrCreate(['customer_id' => Auth::id()]);
-            // Verificar si existe un carrito temporal para el visitante
-            $visitorCart = Cart::where('token', $token)->whereNull('customer_id')->first();
 
-            if ($visitorCart) {
-                // Transferir los elementos del carrito temporal al carrito del usuario registrado
-                foreach ($visitorCart->items as $item) {
-                    // Verificar si el producto ya está en el carrito del usuario
-                    $existingItem = $cart->items()->where('product_id', $item->product_id)->first();
-                    if ($existingItem) {
-                        // Incrementar la cantidad si el producto ya está en el carrito del usuario
-                        $existingItem->quantity += $item->quantity;
-                        $existingItem->save();
-                    } else {
-                        // Agregar el producto al carrito del usuario
-                        $cartItem = new CartItem(['product_id' => $item->product_id, 'quantity' => $item->quantity]);
-                        $cart->items()->save($cartItem);
+            // Verificar si existe un carrito temporal para el visitante
+            $token = request()->cookie('cart_token');
+            if ($token) {
+                $visitorCart = Cart::where('token', $token)->first();
+                if ($visitorCart) {
+                    // Transferir los elementos del carrito temporal al carrito del usuario registrado
+                    foreach ($visitorCart->items as $item) {
+                        $existingItem = $cart->items()->where('product_id', $item->product_id)->first();
+                        if ($existingItem) {
+                            // Incrementar la cantidad si el producto ya está en el carrito del usuario
+                            $existingItem->quantity += $item->quantity;
+                            $existingItem->save();
+                        } else {
+                            // Agregar el producto al carrito del usuario
+                            $cartItem = new CartItem(['product_id' => $item->product_id, 'quantity' => $item->quantity]);
+                            $cart->items()->save($cartItem);
+                        }
                     }
+
+                    // Eliminar el carrito temporal del visitante
+                    $visitorCart->items()->delete();
+                    $visitorCart->delete();
                 }
-                // Eliminar el carrito temporal del visitante
-                $visitorCart->items()->delete();
-                $visitorCart->delete();
             }
 
-            // Asegurarse de que el carrito ya no tenga un token de visitante
-            $cart->customer_id = Auth::id();
-            $cart->token = null;
-            $cart->save();
+            return $cart;
         } else {
-            // Si el usuario no está registrado, creamos o buscamos un carrito temporal por token
+            // Usuario no autenticado: buscar o crear un carrito temporal usando un token de cookie
+            $token = request()->cookie('cart_token') ?: Str::uuid();
             $cart = Cart::firstOrCreate(['token' => $token]);
-            cookie()->queue(cookie('cart_token', $token, 60 * 24 * 30)); // Guardamos el token del carrito en la cookie
-        }
 
-        // Asegurarse de que el carrito ha sido creado y tiene un ID válido
-        if (!$cart->exists) {
-            $cart->save();
+            // Guardar el token del carrito en una cookie
+            cookie()->queue(cookie('cart_token', $token, 60 * 24 * 30)); // 30 días
+            return $cart;
         }
-        return $cart;
     }
 }
